+++ cursus/00-libft/Part1/ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: dvandenb <dvandenb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/28 17:20:16 by dvandenb          #+#    #+#             */
-/*   Updated: 2023/06/28 19:56:43 by dvandenb         ###   ########.fr       */
+/*   Updated: 2023/10/09 15:10:01 by dvandenb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dst, const void *src, long unsigned int n)
+#include <stdlib.h>
+
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char		*to;
 	const char	*from;
