+++ cursus/00-libft/Part1/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: dvandenb <dvandenb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/28 19:53:04 by dvandenb          #+#    #+#             */
-/*   Updated: 2023/06/28 19:53:14 by dvandenb         ###   ########.fr       */
+/*   Updated: 2023/10/09 15:11:35 by dvandenb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long unsigned int	ft_strlen(const char *s)
+#include <stdlib.h>
+
+size_t	ft_strlen(const char *s)
 {
 	int	i;
